[D2] Title: 중간 평균값 구하기, Time: 132 ms, Memory: 21,260 KB -BaekjoonHub

+++ SWEA/D2/1984. 중간 평균값 구하기/중간 평균값 구하기.java
@@ -12,17 +12,17 @@ class Solution
 		{
             int min = Integer.MAX_VALUE;
             int max = Integer.MIN_VALUE;
-            double sum = 0;
+            int sum = 0;
             for(int i = 0; i < 10; i++){
                 int tmp = sc.nextInt();
+                
             	min = Math.min(min, tmp);
                 max = Math.max(max, tmp);
-                
                 sum += tmp;
             }
             
             sum -= (max + min);
-            int avg = (int)Math.round(sum / 8);
+            int avg = (int)Math.round(sum / 8.0);
             System.out.printf("#%d %d\n",tc,avg);
 		}
 	}
